backend/requirements.txt
django
djangorestframework
psycopg2-binary
djangorestframework-simplejwt
python-dotenv
pytz
django-cors-headers
django-filter